refactor(config): split export out of the transfer apply module
apply.rs sat exactly on the 150-line budget, leaving no room for a new
portable preference. The two directions are independent anyway: only
import has a rule to enforce.
